Add modifiers koan, and nix shell definition.
The Nix shell definition should make it a little easier to get started, provided
that the user has Nix installed. At least this way, the user can easily run the
contemplate script on any changes to the koans.

I've only included a few modifiers here, to focus more on the underlying
concepts, and allow for more repetition.