Add koan on combinators.
I'm unsure what order these should be introduced in, or if we should
introduce more than a couple here.  I think that since this koan is about
combinators, we should focus more on the topic of combinators, than on
specific ones.  Later, I can add some konas for specific modifiers.